[stub mirror] chore: refactor code for better readability and some cleanup (#61)
This stuff has breaking changes in the c interface. And i most likely will change more things in the future. But in the end it should lead to a better result.

Take a look at the header for a detailed overview of the changes. Should be straight forward to adapt